FinishUp
we have used reflection to grab information from assemblies that have not been referenced. to test the functionality using reflect, it is bad that we need to handle errors at  runtime, but it shows us how to link the attributes and refelction together to test functionality at run time. So the reflection provides objects and types describe assemblies, modules and types. we can also create instance of the type and modify or access its property, fields and methods.
enables us to access attributes. whhich can be then used to test assemblies in the same directory with the attribute constrains to test its functionality. which is a late binding technology to test functioinbality without referencing at compile time. early binding is referencing.
refecltion can be used for developers to create code for dynamically generating methods and even assemblies at runtime.